requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
python-dotenv==1.0.1
python-multipart==0.0.20
sqlalchemy==2.0.37
sqlalchemy-orm
SQLAlchemy==2.0.37
paramiko==3.4.0
pydantic-settings
apscheduler
fastapi-cache2
psycopg2-binary==2.9.10
pandas
boto3